sensord: change int to intptr_t for avoiding build-break

+++ src/gravity/gravity_sensor.cpp
@@ -129,7 +129,7 @@ bool gravity_sensor::on_start(void)
 	AUTOLOCK(m_mutex);
 
 	m_orientation_sensor->add_client(ORIENTATION_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_orientation_sensor->add_interval((int)this, (m_interval/MS_TO_US), true);
+	m_orientation_sensor->add_interval((intptr_t)this, (m_interval/MS_TO_US), true);
 	m_orientation_sensor->start();
 
 	activate();
@@ -141,7 +141,7 @@ bool gravity_sensor::on_stop(void)
 	AUTOLOCK(m_mutex);
 
 	m_orientation_sensor->delete_client(ORIENTATION_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_orientation_sensor->delete_interval((int)this, true);
+	m_orientation_sensor->delete_interval((intptr_t)this, true);
 	m_orientation_sensor->stop();
 
 	deactivate();
